wimax_manager: fix sign-compare warning.
This fixes the issue of comparison between signed and unsigned
integer expressions.

BUG=chromium:392924
TEST=emerge-daisy platform2 with gcc 4.9, the warning is gone.

Reviewed-on: https://chromium-review.googlesource.com/207383
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>

+++ wimax_manager/gdm_driver.cc
@@ -161,12 +161,12 @@ bool GdmDriver::Initialize() {
                     sizeof(api_param.nonvolatile_dir),
                     "%s",
                     kNonVolatileDirectory),
-           sizeof(api_param.nonvolatile_dir));
+           static_cast<int>(sizeof(api_param.nonvolatile_dir)));
   CHECK_LT(snprintf(api_param.log_path,
                     sizeof(api_param.log_path),
                     "%s",
                     kLogDirectory),
-           sizeof(api_param.log_path));
+           static_cast<int>(sizeof(api_param.log_path)));
   api_param.log_level = 1;
   GCT_API_RET ret =
       GAPI_Initialize(GCT_WIMAX_SDK_EMBEDDED_EAP_ENABLED, &api_param);
